backend/requirements.txt
fastapi
torch
onnx
onnxruntime
numpy
pillow
flask
flask-cors
requests
gunicorn
uvicorn